requirements.txt
alembic==1.4.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
webpay-bahamta==0.1.1
Werkzeug==1.0.1
